LPS-55479 Move dependent classes to portal-test-internal
