+++ portal-test-internal/src/com/liferay/portal/cache/test/TestCacheReplicator.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache;
+package com.liferay.portal.cache.test;
 
 import com.liferay.portal.kernel.cache.CacheReplicator;
